+++ llms/setup.py
@@ -32,7 +32,7 @@ setup(
             "mlx_lm.lora = mlx_lm.lora:main",
             "mlx_lm.merge = mlx_lm.merge:main",
             "mlx_lm.server = mlx_lm.server:main",
-            "mlx_lm.model = mlx_lm.model:main",
+            "mlx_lm.manage = mlx_lm.manage:main",
         ]
     },
 )
